# server/openapi_spec.yaml
openapi: 3.0.0
info:
  title: Astronova Backend API
  description: Comprehensive astrology API with AI-powered insights, birth charts, compatibility matching, and personalized reports
  version: 1.0.0
  contact:
    name: Astronova Support
    email: admin@100xai.engineering
servers:
  - url: https://astronova-ghcr.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: "Admin API token. Production admin endpoints also accept Authorization: Bearer <ADMIN_API_TOKEN>."
  schemas:
...
    BirthData:
      type: object
      required:
        - date
        - time
        - timezone
        - latitude
        - longitude
      properties:
        date:
          type: string
          format: date
          example: "1990-05-15"
        time:
          type: string
          format: time
          example: "14:30"
        timezone:
          type: string
          example: "America/New_York"
        latitude:
          type: number
          format: float
          example: 40.7128
        longitude:
          type: number
          format: float
          example: -74.0060
...
paths:
...
  /api/v1/reports:
    get:
      summary: Get reports service info
      tags:
        - Reports
      responses:
        '200':
          description: Service information
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                  status:
                    type: string
                  endpoints:
                    type: object
    post:
      summary: Generate astrological report (alias)
      tags:
        - Reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportType:
                  type: string
                  enum: [birth_chart, career_forecast, love_forecast, year_ahead]
                birthData:
                  $ref: '#/components/schemas/BirthData'
                userId:
                  type: string
      responses:
        '200':
          description: Generated report
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                  type:
                    type: string
                  title:
                    type: string
                  summary:
                    type: string
                  keyInsights:
                    type: array
                    items:
                      type: string
                  downloadUrl:
                    type: string
                  generatedAt:
                    type: string
                    format: date-time
                  status:
                    type: string
  /api/v1/reports/full:
    post:
      summary: Generate comprehensive astrological report
      tags:
        - Reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reportType
              properties:
                reportType:
                  type: string
                  enum: [birth_chart, career_forecast, love_forecast, year_ahead]
                birthData:
                  $ref: '#/components/schemas/BirthData'
                userId:
                  type: string
                options:
                  type: object
      responses:
        '200':
          description: Generated report
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                  type:
                    type: string
                  title:
                    type: string
                  summary:
                    type: string
                  keyInsights:
                    type: array
                    items:
                      type: string
                  downloadUrl:
                    type: string
                  generatedAt:
                    type: string
                    format: date-time
                  status:
                    type: string
...
  /api/v1/reports/generate:
    post:
      summary: Generate report (alias)
      tags:
        - Reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportType:
                  type: string
                  enum: [birth_chart, career_forecast, love_forecast, year_ahead]
                birthData:
                  $ref: '#/components/schemas/BirthData'
                userId:
                  type: string
      responses:
        '200':
          description: Generated report
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                  type:
                    type: string
                  title:
                    type: string
                  summary:
                    type: string
                  keyInsights:
                    type: array
                    items:
                      type: string
                  downloadUrl:
                    type: string
                  generatedAt:
                    type: string
                    format: date-time
                  status:
                    type: string
  /api/v1/reports/user/{user_id}:
    get:
      summary: List reports for a user
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          description: Maximum number of recent reports to return. Defaults to 25 and caps at 50.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 25
      responses:
        '200':
          description: Report list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    reportId:
                      type: string
                    type:
                      type: string
                    title:
                      type: string
                    content:
                      type: string
                      description: Summary preview for list views. Full report content is available through report detail/PDF surfaces.
                    summary:
                      type: string
                    keyInsights:
                      type: array
                      items:
                        type: string
                    downloadUrl:
                      type: string
                    generatedAt:
                      type: string
                      format: date-time
                    userId:
                      type: string
                    status:
                      type: string
        '401':
          description: Authorization required
        '403':
          description: Authenticated user does not match requested user
  /api/v1/reports/{report_id}/pdf:
    get:
      summary: Download report PDF
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: report_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF bytes
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Authorization required for persisted reports
        '403':
          description: Authenticated user does not own the report
...
  /api/v1/numerology/report:
    post:
      summary: Generate Loshu numerology report
      tags:
        - Numerology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dob
              properties:
                dob:
                  type: string
                  example: "1999-12-24"
                phone_digit_sum:
                  type: integer
      responses:
        '200':
          description: Numerology report generated
        '400':
          description: Invalid date of birth or payload
  /api/v1/rajayoga/report:
    post:
      summary: Generate full Raja Yoga report
      tags:
        - Raja Yoga
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - planet_data
                - lagna
              properties:
                planet_data:
                  type: object
                lagna:
                  type: string
      responses:
        '200':
          description: Raja Yoga report generated
        '400':
          description: Invalid chart payload
...
  /api/v1/report/{path}:
    get:
      summary: Legacy report endpoint (redirect)
      tags:
        - Reports
      deprecated: true
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '307':
          description: Redirect to /api/v1/reports
    post:
      summary: Legacy report endpoint (redirect)
      tags:
        - Reports
      deprecated: true
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '307':
          description: Redirect to /api/v1/reports
    put:
      summary: Legacy report endpoint (redirect)
      tags:
        - Reports
      deprecated: true
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '307':
          description: Redirect to /api/v1/reports
    delete:
      summary: Legacy report endpoint (redirect)
      tags:
        - Reports
      deprecated: true
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '307':
          description: Redirect to /api/v1/reports
    patch:
      summary: Legacy report endpoint (redirect)
      tags:
        - Reports
      deprecated: true
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '307':
          description: Redirect to /api/v1/reports
...
  /api/v1/reports/{report_id}/status:
    get:
      summary: Get report generation status
      tags:
        - Reports
      parameters:
        - in: path
          name: report_id
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Report status
...